[native] Extract AppNavigator, RootNavigator, and RootRouter
